refactore: Use separated base classes for version references and non version references.
On objectmap level we need to distinguish if a reference points to an
old version or is an normal "data" reference between isheets. To use the
the IVersionable isheets causes circle import and increase SRP.
Instead we use sparated base classes for theses references.

Note: the hole reference system is to comlex and has no good Separation
of Concern.