scss lint: don't enforce single-line comments
Hologram currently requires multi-line comments for its inline
documentation. In order to not fail, the scss comment linter is now
disabled.

Note that multi-line comments are not stripped from the compiled CSS
code, which is not nice. There's an issue demanding the usage of single-
line comments for hologram documentation in the hologram tracker [0]. We
should follow the progress there.

[0] https://github.com/trulia/hologram/issues/110