do not depend on specific user service name
before this, the login directive used the name of the user service
(adhUser) which was defined in a completely different place.

To avoid depending on a specific name we now pass the names of both the
service and the directive to a new `register` function which registers
them to our application in one step.