[R] dropped undefensible claims on generic adapter classes
(it might be a good idea to do it.  i think the problem was that we
tried to access the concrete type of a type parameter at compile-type,
but the concrete type is not available at all (at compile-time, it
depends on run-time context, and at run-time, it has been compiled
away).  the fact that adapter classes were generic may have been
incidental.