fix check_code return code
check_code did implicitly return the return code of the last command.
As check_code is used in a pre commit hook it was possible to commit
changes that broke one or more tests as long as the final test was
successful.

This fixes that bug by summing up all return codes of the individual
tests.