Introduce separate identifier declarations in module head
The previous commit did not solve anything, but just did a rewind to
the beginning of the discussion.  The problem is that the tslint rule
label-undefined does understand lambdas and thus warns about the
following code:

  var f1 = () => f2();
  var f2 = () => return;

Worse, it warns about this:

  var f1 = () => f2();
  var f2 = () => f1();

In both cases, it would be perfectly possible to statically reason
that no problem exists.  (In the second, there is a termination issue,
which in this simple case is obvious, but the halting problem states
that termination issues are never the business of the compiler.)

Options:

 1. disable 'label-undefined' in etc/tslint.json.

 2. require use of function over =>.  (this is what happened in the
    previous commit.)

 3. re-arrange functions so that calls always go back in the
    declaration sequence.

 4. separate declarations and definitions.

I am unhappy with all three, but you can only get so far without
changing to a better language.

(2) and (3) are unacceptable.  (2) re-introduces the broken javascript
'function' semantics that we wanted to get rid of.  (3) not only makes
it much harder for the programmer to tell a coherent story, but it
does not solve the problem for recursive functions.

(1) is acceptable, but I'm not sure what valid warnings we'd be giving
up.

(4) is my clear preference, and I have changed Http.ts so that it
follows that rule.  Declarations do not **have to** be separated from
definitions, but they **can**.