use JsApiReporter instead of ConsoleReporter
AFAIK ConsoleReporter is not deprecated, but it is not included in the
jasmine distribution we use. jasmine_egg (the python integration
provided by the jasmine project) uses the JsApiReporter. So I suggest we
do the same.

You can now access the results of the test run via the global
jsApiReporter variable.

This also fixes the spec filter.