Track listing elements by path instead of $index
This results in a constant mapping between element paths and directives.

This commit also reverts 52db3e654d8808833134c1a6d2afae59738b14a0 (watch
element.path), as we simply don't change the path in an element anymore.

This assumes that there are no duplicate listing elements (which is
currently the case). We should take care of that possible case somehow,
a FIXME is added.