Fix proposal creator
As scope.content is loaded asynchronously the creator is not available
when the template is first rendered. One option to fix this would be to
$watch the path attribute.

I tried another approach here: The template is only created (ng-if) when
content is loaded. This potentially results in much less async in the
child directives.