Implement Batch API FE
Some of this was already done in the past. However, we could not yet
quite decide on a frontend API. With this commit I dropped the idea of
having a adhHttp-like transaction object completely and concentrated on
having an easy to use API instead.

The main problem was that we did not know what get/put/post should
return (promise to the real response vs. preliminary data). On the one
hand we want to get the real response. On the other hand we need
preliminary data during the request.

In the end I went for the preliminary data. This data is returned
syncronousely (no promise) and is a Content-like object which may
contain additional data. `first_version_path` is a current example,
but others may be added in the future.

Still some responsability is left to the user. They must take care of
only using preliminary data within the transaction themselves. But it
is no longer possible to create deadlocks due to promises that are only
resolved after the transaction has been committed. The transaction also
throws errors if you try to use it after the commit.