fix comment editing
Before the previous commit, scope.path was only used to initialize the
directive. After that, versionPromise was used to keep track of the
currently displayed version.

In the previous commit, scope.path was used as the poolPath argument to
the inner create form. This led to some strange behavior.

So this commit fixes all that by updating scope.path on each call to
updateScope().