listen to 'storage' events and update login status
When multiple instances of adhocracy are embedded on a single host page,
the login status on all of them should be in sync. This also stretches
to other tabs and windows.

The user token is stored in local storage, so I implemented this
behavior by listening to 'storage' DOM events and updateing the user
service accordingly.