Refactor: fix dep injections
The relationship between loginDirective and the (inlined) corresponding
loginController was a bit confused: $location was injected into the directive,
but then used only in the controller.  To make things a little more transparent,
there is an extra export loginController now that takes all its dependencies
(including $location), and that is called from loginDirective.

Same for registerDirective.