API change: resource references are URLs instead of paths; refactor adhocrcy.schema
Use resource URLs instead of resource path:

This makes it possible split the "frontend" server (static files and
javascript) and the backend server (rest api). Also it is more easy to
deploy the application with virtual hosting (for example add prefix to
the url path).

Refactor Refernce SchemaNodes in adhocracy.Schema and rename
ListOfUniqueReferences to UniqueRefernces.