Revert "Add acceptance test creating broken database"
This reverts commit 2560bf6c969c0644fef78eaaaae17c3e4aa8bca0. This
regression test has done its duty, as the bug is fixed in
e9b6b0c74341f8d53638fcff4f04d34ef9010196, but it doesn't work anymore,
as branch 2014-06-joka-default-permissions-and-roles (#125) disabled the
"create proposal" button for anonymous users.

So this test is removed again. It could be rewritten as a functional
backend test.