Fix: buildout, Makefile
tslint_check_adhocracy was only running inside the current directory.
This does not work as expected in parts/statis/js/, so i switched it
to always running in $A3_ROOT.  which is now also exported as a shell
variable in source_env.  which is used in Makefile instead of pathless
commands for transparency and robustness.