Fix: error in using typescript.
cc, not sure if i understood your code, but i think you wanted to say
that function argument fieldName is to have type string.  what you did
was say that there is a function argument fieldName of any type, and
a local variable of the same name that has type string.  not sure why
this has produced the error it has, but at least it did produce an
error.  :)