Do not encode date objects as strings (fixes #458)
For some time now, date fields in resources are automatically
transformed to JavaScript Date objects.

The adh-time directice than takes this Date as a DOM-attribute (@). So
it is converted to a string before. Apart from this being unecessary,
the resulting string can not parsed as a date.

My fix is to switch the attribute on adh-time to biderictional binding
(=) so the Date object can be passed as-is.