update ruby gems
Note that `bin/gem outdated` reports a new version of compass, but the
changelog does not mention it.  I looked at the commits and decided not
to do this update just now.

I did not see any notable changes except for the new hologram version
which seems to have some new options that we should look into.

Changelogs:

-   http://sass-lang.com/documentation/file.SASS_CHANGELOG.html
-   https://github.com/causes/scss-lint/blob/master/CHANGELOG.md
-   http://susydocs.oddbird.net/en/latest/changelog/
-   https://github.com/trulia/hologram/blob/master/CHANGELOG.md
-   http://compass-style.org/CHANGELOG/