Change annotated-section implementation
Initially, this was implemented with two floating boxes and a
`overflow: hidden` clearfix.

In 5c4d892 the clearfix was changed to `pie-clearfix` which uses an
`::after` pseudo-element with `clear: both` in order to not clip
outlines.

Unfortunately, the second version has proplems with sticky positioning
in firefox (the sticky navigation floats and so the clearfix tries to
clear that, too).

This newest version uses `display: inline-block` instead of float. This
would be a nice solution if it were not for the whitespace between the two
blocks. Two remove that additional space, I set `font-size: 0` in the
container and `font-size: $font-size-normal` in the children.