rename eventHandler to eventManager
the generic eventManager was introduced a long time ago to allow service
to expose an interface similar to jquery's on/off/trigger.  For some
reason, it was called "eventHandler". But this term is already used for
callbacks that are registered fir an event. So I propose to rename it to
"eventManager".