Skip Mercator tests on travis because they don't work on sauce labs
The sauce labs browser doesn't have the local files on the travis server
available. This problem is reported in angular/protractor#612 and
addressed in SeleniumHQ/selenium#219 using the setFileDetector API - the
next version of protractor will support this.

The proposed workaround to use a URL instead of a path in `sendKeys` as
proposed in the protractor issue doesn't work (seems to work only on
windows).