allow to add transitions to tabset-panes height
CSS transitions on height to not work if the height is set to "auto".
So with this the height of tabset-panes is explicitly set to the childs
height.

Note that this will have unexpected results if you apply a padding to
tabset-panes.

To use this, use some CSS like the following:

    .tabset-panes {
        transition: height 0.8s;
        overflow: hidden;
    }