Allow to disable proposal creation in frontend
This is done through a configuration option out of simplicity.

It would be better to make use of permissions for that, however this
doesn't work, as we also want to show the button if the user isn't
logged in, but login would help.

In order to do this we need to add something like "login-would-help" to
the OPTIONS request.