Fix anonymous cannot create meinverlin proposal becaus editing in disallowed
To create a proper propsal you need the create permission to create
version 0 but also the edit permission to create version 1 (with
concrete data). Becaucse editing is dissalowed the frontend cannot crate
a proper poposal.

This hack adds an workaround. If only only one versions (version 0)
existis the permission check for adding new resources is ignored.
If more versions exist the permission check is reanabled.