Move mein berlin specific code from core to meinberlin (fixes #1028)
The reason this was here in the first place is that we had trouble with
nested injection (`adhMapListing` and `adhMapListingInternal`).  But when I
tried this now it just worked.

I also tried to reproduce the original issue by creating a minimal
application: http://plnkr.co/edit/YhwT9X1ra4R6E7PeDEnF . I was not able
to reproduce the issue, even with older angular versions.