Add title sheet for badges in import_users script
Set title sheet when assigning a badge to an user. The title of the
badge is name given in the JSON file, the name in the graph and url is
the lower case version of the title.

No further validation is done yet as we don't have a good concept to
validate input in the scripts.