raise tab width calculation precision
the tab width is measured in percentage. Thie issue with this is that
the rounding that occours when converting % into px may leave us with
more than 100%.

To counter that, we floor the width. However, 1% is actually a lot. So
this improves the precision slightly.

This may be supplemented by some CSS hacks. But it already improves the
situation a lot.