Change default autobahn port to 6561
Port 8080 is frequently used by other tools (e.g. tomcat, which is
installed in the policycompass project) and may conflict with autobahn.

The intention of using 8080 for autobahn is that this is on some
firewalls not blocked, but this doesn't matter for us anyway as we
typically put nginx in frontend of autobahn and use a dedicated path
/websockets for autobahn.

Port 6561 is somewhat inline with 6551 and 6541.