Switching the openly bound isImageMissing property from the user image and image directive to a callback instead for better isolation.
Rationale: the Image directive already shows it's fallback image until it could load the real image. Thus if you never give it a path to work with or one that doesn't contain an image, it doesn't need to watch the isImageMissing property instead.

The user profile image directive is easier as it just messages it's callback if the image couldn't be found and hands through callbacks it receives from the image directive.

The user profile image editor doesn't need anything besides the callbacks as it doesn't need to communicate 'inwards' towards the image tag.