Temporary fix for private processes
Fixes only 'participate' state. Private processes are broken since
'anonymous' role is not used anymore.

All participants have both the 'participant' and 'everyone' roles, so
setting 'Deny' for 'view' permission for 'everyone' result in a 'Deny',
if participant does not have explicity 'Allow' in the private workflow.
Setting explicitely 'Allow' do not allow us to us inheritance with the ~
symbol.