allow to set worflow state to current state
The workflow sheet only allows to put request with the next workflow
state or without workflow state data. "Resetting" the exiting state is
not possible. This is not the expected behaviour (for all other sheets
you can make put request that do not change the data) and does not work
with the Substanced Admin Interface.

This commit changes the state validator. The current state is now a
valid value. If you set the current state again a state transition is
not triggered.