remove text-align mixin
its was meant to ployfill `text-align: start/end`, but in order for this
to be effective we would also need to be able to set something like
`float: start` or `margin-end: …`. This issue is addressed in
https://www.w3.org/TR/css-writing-modes-3/

For now a better approach to bidirectional webdesign is to have a
separate stylesheet where `left` and `right` are swapped.