Add roadmap to documentation
We have been working on a project roadmap internally and now we want to
publish it.

We also discussed publishing it as github milestones or on a github wiki
page but finally decided that it best fits in our documentation.

For anyone who has seen the previous (internal) draft: I stripped out
anything that I thought should not be publizised and also anything that
is not directly connected to the software but rather to individual
deployments or organizational aspects.