Fix regression: blob storage path is not package specific for S1
The blob storage path for the ZEO client (pyramid app) was set to
var/blobs, the path for the zeo server instead to
var/db/<package>/blobs.

This was causing file not found errors.

Now both path are pointing to var/db/<package>/blobs for the s1 package.
The other package still need to be fixed.